# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
      "DevLoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "email"
        ],
        "title": "DevLoginRequest",
        "type": "object"
      },
      "DocumentCompleteOut": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentOut"
          },
          "pipeline_run_id": {
            "format": "uuid",
            "title": "Pipeline Run Id",
            "type": "string"
          }
        },
        "required": [
          "document",
          "pipeline_run_id"
        ],
        "title": "DocumentCompleteOut",
        "type": "object"
      },
      "DocumentCreate": {
        "properties": {
          "filename": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "mime": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Mime",
            "type": "string"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "mime",
          "size_bytes"
        ],
        "title": "DocumentCreate",
        "type": "object"
      },
      "DocumentListItem": {
        "properties": {
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Id"
          },
          "latest_run_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "filename",
          "mime",
          "size_bytes",
          "content_hash",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentListItem",
        "type": "object"
      },
      "DocumentOut": {
        "properties": {
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "filename",
          "mime",
          "size_bytes",
          "content_hash",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
      "DocumentStatus": {
        "enum": [
          "pending_upload",
          "uploaded",
          "processing",
          "ready",
          "failed"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentUploadOut": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentOut"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "document",
          "upload_url"
        ],
        "title": "DocumentUploadOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MeResponse": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "orgs": {
            "items": {
              "$ref": "#/components/schemas/OrgSummary"
            },
            "title": "Orgs",
            "type": "array"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "orgs"
        ],
        "title": "MeResponse",
        "type": "object"
      },
      "OrgSummary": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "role"
        ],
        "title": "OrgSummary",
        "type": "object"
      },
...
      "RunStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Transformation Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/auth/dev-login": {
      "post": {
        "description": "Password-less login for local development.\n\nRegistered only when auth_mode=dev; config validation refuses that mode\nin production (app.config).",
        "operationId": "dev_login_v1_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dev Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/me": {
      "get": {
        "operationId": "me_v1_auth_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "atc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Atc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/workos/callback": {
      "get": {
        "operationId": "workos_callback_v1_auth_workos_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workos Callback",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/orgs/{org_id}/documents/{document_id}/complete": {
      "post": {
        "description": "Step 2: confirm the blob landed, then kick off the ingestion pipeline.",
        "operationId": "complete_document_upload_v1_orgs__org_id__documents__document_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "atc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Atc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentCompleteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Document Upload",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/v1/orgs/{org_id}/projects/{project_id}/documents": {
      "get": {
        "operationId": "list_documents_v1_orgs__org_id__projects__project_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "atc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Atc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentListItem"
                  },
                  "title": "Response List Documents V1 Orgs  Org Id  Projects  Project Id  Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "description": "Step 1 of upload: register the document, hand back a presigned PUT URL.\n\nThe browser uploads directly to object storage \u2014 document bytes never\nstream through the API process.",
        "operationId": "create_document_upload_v1_orgs__org_id__projects__project_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "atc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Atc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Document Upload",
        "tags": [
          "documents"
        ]
      }
    }
  }
}